rename_commit.rb: clarify meaning of 'first' commit
The wording of the challenge suggests that one needs to amend the commit
that one would get from the following command:

```console
$ git rev-list HEAD | tail -1
```

i.e. the first commit which was made. The "first" commit this level
talks about is, to my mind, the second commit.

After a degree of Googling, I found the magic incantation to amend the
root commit. This is non-trivial with ``git rebase``. Admittedly I
should've checked which commit actually *had* the typo but this problem
would've been avoided all together had the challenge been clearer.