Correct order of commits in "reorder" solution
The solution incorrectly expects "Third commit" to be the root commit,
"Second commit" to be the second commit, and so on.

These changes correct the order. The solution now expects that "Initial
Setup" is the root commit, that "First commit" is the second commit, and
so on.

Many thanks to Gary for finding a bug in my original implementation and
suggesting a better one. He deserves the credit here. :-)