flesh out game instructions in README
The instructions for how to use Githug assumed a lot of knowledge that
many beginners do not have (Ruby, filesystem idioms, permissions).

This makes the Ruby prerequisite explicit (not everyone who
wants to learn git is learning Ruby).

Since people who are learning git without ruby might already have Ruby
installed at the system-level, this adds a note about running `gem
install` with `sudo`.

It also separates out "how to play" from the installation instructions,
and removed the `test` command from the list of commands, since that
is contributor-related and not relevant for people who are just using
the tool to learn how to use git. The `test` command is documented
below in the contribution section.